move button for exporting script to builder panel
